Fix review comments: * Rename note_position -> with_position * Use Peekable<Fuse<I>> in WithPosition * Forward size_hint from WithPosition to inner iterator * Use derive(Copy, Clone, Debug, PartialEq) on Position * Change next() impl conditionals * Fix import ordering * Fix line length
