+++ src/lib.rs
@@ -70,12 +70,12 @@ pub mod structs {
     pub use intersperse::Intersperse;
     pub use kmerge_impl::KMerge;
     pub use pad_tail::PadUsing;
-    pub use note_position::NotePosition;
     pub use rciter_impl::RcIter;
     pub use repeatn::RepeatN;
     pub use sources::{RepeatCall, Unfold, Iterate};
     pub use tee::Tee;
     pub use tuple_impl::{TupleBuffer, TupleWindows, Tuples};
+    pub use with_position::WithPosition;
     pub use zip_eq_impl::ZipEq;
     pub use zip_longest::ZipLongest;
     pub use ziptuple::Zip;
@@ -85,9 +85,9 @@ pub use cons_tuples_impl::cons_tuples;
 pub use diff::diff_with;
 pub use diff::Diff;
 pub use minmax::MinMaxResult;
-pub use note_position::Position;
 pub use repeatn::repeat_n;
 pub use sources::{repeat_call, unfold, iterate};
+pub use with_position::Position;
 pub use zip_longest::EitherOrBoth;
 pub use ziptuple::multizip;
 mod adaptors;
@@ -102,7 +102,6 @@ mod groupbylazy;
 mod intersperse;
 mod kmerge_impl;
 mod minmax;
-mod note_position;
 mod pad_tail;
 mod rciter_impl;
 mod repeatn;
@@ -110,6 +109,7 @@ mod size_hint;
 mod sources;
 mod tee;
 mod tuple_impl;
+mod with_position;
 mod zip_eq_impl;
 mod zip_longest;
 mod ziptuple;
@@ -906,16 +906,20 @@ pub trait Itertools : Iterator {
     /// ```
     /// use itertools::{Itertools, Position};
     ///
-    /// let it = (0..4).note_position();
-    /// itertools::assert_equal(it, vec![Position::First(0), Position::Middle(1), Position::Middle(2), Position::Last(3)]);
+    /// let it = (0..4).with_position();
+    /// itertools::assert_equal(it,
+    ///                         vec![Position::First(0),
+    ///                              Position::Middle(1),
+    ///                              Position::Middle(2),
+    ///                              Position::Last(3)]);
     ///
-    /// let it = (0..1).note_position();
+    /// let it = (0..1).with_position();
     /// itertools::assert_equal(it, vec![Position::Only(0)]);
     /// ```
-    fn note_position(self) -> NotePosition<Self>
+    fn with_position(self) -> WithPosition<Self>
         where Self: Sized,
     {
-        note_position::note_position(self)
+        with_position::with_position(self)
     }
 
     // non-adaptor methods
